some small changes, and beginning of chest interaction

+++ Protocol/play_serverbound.go
@@ -312,7 +312,7 @@ type PlayerBlockPlacement struct {
 	Location                  lib.Position
 	Face                      lib.VarInt
 	Hand                      lib.VarInt
-	CursorX, CursorY, CursorZ byte
+	CursorX, CursorY, CursorZ float32
 }
 
 // UseItem is sent when the client tries to use an item.
